requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
MarkupSafe==3.0.3
narwhals==2.13.0
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.52.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
